# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a Test Calculator. For Transfer To.",
    "version": "1.0.0",
    "title": "MegaSoft Calculator",
    "contact": {
      "email": "abdullahrahim87@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/calc",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/factorials": {
      "post": {
        "tags": [
          "Factorials"
        ],
        "summary": "Factorials",
        "description": "",
        "operationId": "Factorials",
        "consumes":["application/x-www-form-urlencoded"],
        "produces": [
          "application/json"
        ],
        "parameters" : [
          {
            "name": "number",
            "in": "formData",
            "description": "Number",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
...
}